Use requirements-prod.txt to exclude heavy ML libs in production

+++ build.sh
@@ -2,8 +2,12 @@
 # Thoát ngay nếu có lỗi
 set -o errexit
 
-# Cài đặt thư viện
-pip install -r requirements.txt
+# Cài đặt thư viện (production dùng file nhẹ hơn)
+if [ -f requirements-prod.txt ]; then
+    pip install -r requirements-prod.txt
+else
+    pip install -r requirements.txt
+fi
 
 # Gom các file static (CSS, JS, Images) để Render phục vụ
 python manage.py collectstatic --no-input
